// lib/_components/button/types.ts
import { SyntheticEvent } from 'react';

export type ButtonClickEventType = {
  /** ивент событие */
  event: SyntheticEvent<HTMLButtonElement>;
};

export type ButtonSize = 'big' | 'small';

export type ButtonType = 'button' | 'submit' | 'reset';

export type TargetType = '_blank' | '_self' | '_parent' | '_top';

export type ButtonVariant =
  | 'main'
  | 'accent'
  | 'interface'
  | 'success'
  | 'adaptive'
  | 'interface-menu'
  | 'accent-menu'
  | 'remove'
  | 'add'
  | 'link'
  | 'only-icon';

export type DarkThemePresets = 'neutral' | 'positive' | 'negative';